Move the typescript parser outside of jest-editor-support to jest-test-typescript-parser
