+++ packages/jest-editor-support/src/index.js
@@ -15,7 +15,7 @@ const ProjectWorkspace = require('./ProjectWorkspace');
 const Runner = require('./Runner');
 const Settings = require('./Settings');
 const {Expect, ItBlock, Node} = require('./parsers/ParserNodes');
-const {parse} = require('./ScriptParser');
+const {babylonParser: parse} = require('./parsers/BabylonParser');
 const TestReconciler =  require('./TestReconciler');
 
 module.exports = {
